kytesmum: kytesmum_update 93.73->94.27 (the (s16) test-cast re-execution - T keeps the double-extsh dead wrap tests; the import folded them; faithful-bug preservation)

+++ src/main/dll/DR/dll_0266_kytesmum.c
@@ -81,10 +81,10 @@ void kytesmum_update(int obj) {
         }
     }
     diff = (s16)((setup->yaw << 8) - (u16)kytesMum->yaw);
-    if (diff > 0x8000) {
+    if ((s16)diff > 0x8000) {
         diff = (s16)((diff - 0x10000) + 1);
     }
-    if (diff < -0x8000) {
+    if ((s16)diff < -0x8000) {
         diff = (s16)((diff + 0x10000) - 1);
     }
     if (diff != 0) {
